requirements.txt
numpy==1.24.0
pandas==2.2.3
scikit-learn==1.4.2
PyYAML==6.0.3
torch==2.5.1
matplotlib>=3.7
tqdm>=4.66
